requirements.txt
Flask==0.10.1
python-dateutil ==2.6.0
psycopg2==2.7.1